pyproject.toml
[project]
name = "tabdce"
version = "0.1.0"
description = "Counterfactual explainations with diffusion model"
authors = [
    {name = "Jan Maslowski"}
]
readme = "README.md"
requires-python = ">=3.11"

[tool.poetry.dependencies]
python = "^3.11" 
torch = "^2.0.0" 
numpy = "^1.24"
scikit-learn = "^1.3"
pyyaml = "^6.0"
tqdm = "^4.66"
seaborn = "^0.13.2"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "jupyter (>=1.1.1,<2.0.0)",
    "ipykernel (>=7.1.0,<8.0.0)",
    "matplotlib (>=3.10.7,<4.0.0)"
]
